🔥 remove : Delete unuse leg stretching
