+++ src/utils/detect-pose/leg-stretching.js
@@ -5,6 +5,8 @@ import { getKeypointsObject, getAngle } from "../estimate-pose";
 // 추후 함수명은 동작 이름으로 변경. 대문자로 시작.
 // 다리 스트레칭
 export default function LegStretching() {
+  // console.log("leg stretching work");
+
   const [count, setCount] = useState(0);
   const [step, setStep] = useState(0);
 
@@ -12,26 +14,20 @@ export default function LegStretching() {
     if (pose.keypoints[11].score < 0.1 || pose.keypoints[13].score < 0.1) {
       console.log("can not see");
     } else {
-      const { leftHip, rightHip, leftKnee, rightKnee, leftAnkle, rightAnkle } =
-        getKeypointsObject(pose);
+      const { rightShoulder, rightHip, rightKnee } = getKeypointsObject(pose);
 
-      const anglesLegs = {
-        rightHigh: getAngle(rightHip.x, rightHip.y, rightKnee.x, rightKnee.y),
-        rightLow: getAngle(
-          rightKnee.x,
-          rightKnee.y,
-          rightAnkle.x,
-          rightAnkle.y
+      const angles = {
+        upperBody: getAngle(
+          rightHip.x,
+          rightHip.y,
+          rightShoulder.x,
+          rightShoulder.y
         ),
-        leftHigh: getAngle(leftHip.x, leftHip.y, leftKnee.x, leftKnee.y),
-        leftLow: getAngle(leftHip.x, leftHip.y, leftAnkle.x, leftAnkle.y),
+        leg: getAngle(rightHip.x, rightHip.y, rightKnee.x, rightKnee.y),
       };
 
-      const left = checkLeftLegStretching(anglesLegs);
-      const right = checkRightLegStretching(anglesLegs);
-
-      if (step == 0 && left) setCount((count) => count + 1);
-      else if (step == 1 && right) setCount((count) => count + 1);
+      const legStretching = checkLegStretching(angles);
+      if (step == 0 && legStretching) setCount((count) => count + 1);
     }
   });
 
@@ -40,30 +36,15 @@ export default function LegStretching() {
       setStep((step) => 1);
       setCount((count) => 0);
     }
-    if (step == 1 && count > 20) {
-      setStep((step) => 2);
-      setCount((count) => 0);
-    }
   }, [step, count]);
 
   return [count, step, checkPoses];
 }
 
-// 오른쪽 종아리 스트레칭
-// 왼다리 직각으로 굽혀서 앞으로, 오른쪽 다리는 쭉 펴서 뒤로, 왼쪽 보며 실행
-function checkRightLegStretching(anglesLegs) {
-  if (anglesLegs.rightLow < 110) return false;
-  else if (20 > anglesLegs.leftHigh || anglesLegs.leftHigh > 70) return false;
-  else if (80 > anglesLegs.leftLow || anglesLegs.leftLow > 100) return false;
-  else return true;
-}
-
-// 왼쪽 종아리 스트레칭
-// 앞과 반대
-function checkLeftLegStretching(anglesLegs) {
-  if (anglesLegs.leftLow > 70) return false;
-  else if (110 > anglesLegs.rightHigh || anglesLegs.rightHigh > 160)
-    return false;
-  else if (80 > anglesLegs.leftLow || anglesLegs.leftLow > 100) return false;
+// 양 다리 스트레칭
+// 몸 앞으로 숙이기
+function checkLegStretching(angles) {
+  if (70 > angles.leg || angles.leg > 110) return false;
+  else if (0 > angles.upperBody) return false;
   else return true;
 }
